requirements.txt
flask
PyPDF2
python-docx
python-pptx
comtypes
pywin32
pikepdf
Pillow
PyMuPDF
